Fix checkout error due to git lfs on window (MR !1645)
The behavior of the git scm plugin is different wrt LFS checkouts on
windows vs. linux. On windows, the 'LFS pull after checkout' option is
required, and without it jenkins completely fails to check out the repo.
On linux, the same option causes multibranch pipeline merges (such as
those used for PR branches) to fail with conflicts.

This will conditionally use the option in our only cross-platform
Jenkins pipeline.

https://issues.jenkins-ci.org/browse/JENKINS-52935

Approved-by: Chris Bayruns