Merged in pb/osf-v10-destagger-fix (MR !1643)
Fix OSF V10 destaggering due to changes in px_offsets

Since 1.14 we changed px_offsets and staggering/destaggering
directions accordingly. Thus old drives on wardens (still recorded in
OSF V10) destaggered PNG encoded lidar_scans by shifting pixels to
one direction with not-reversed offsets (for example: (0, 6, 12, 18) for
1024 mode) but recovery happens with changed reversed px_offsets
((18, 12, 6, 0) for 1024 mode) and shift again to the same direction.

Which leads to all columns be shifted to the right by 18 pixels
(for 1024 mode).

So we are applying the post reconstruction fix for lidar scans recovered
from OSF V10 and shifting back columns to the full cycle (e.g. 18 pixels
for 1024 mode)

Add forced lidar frame re-coding in frameChunker for OSF V10 so to
convert lidar_scans and encode them to the correct direction.

Approved-by: Tomas Svarovsky <tomas.svarovsky@ouster.io>
Approved-by: Dima Garbuzov <dima.garbuzov@ouster.io>