Use setuptools for ouster-sensor-tools package (MR !1652)
Since external teams are using the checker tool in this package, it
makes sense to use the most "standard" packaging tools and not require
them to learn poetry.

Also specify a lower bound for the dataclasses backport to work around
the issue linked below. The package should be installable by pip and
poetry in environments with both python 3.6 and 3.7.

https://github.com/python-poetry/poetry/issues/1413

Approved-by: Tomas Svarovsky
Approved-by: Pavlo Bashmakov