Batch scans in python without holding GIL (MR !1674)
Proof of concept for collecting sensor data without dropping packets in
Python. Due to the GIL and high data rate of 128-beam sensors, it's
really hard to read packets while doing any other work. This PR exposes
a easy way to create a separate thread dedicated to collecting sensor
data, that only needs to synchronize with the python interpreter when
handing off a pointer into a (thread-safe) queue.

Approved-by: Chris Bayruns
Approved-by: Pavlo Bashmakov