Add client version from cmake to metadata json (MR !1709)
Instead of pulling the version number from the git tag, use the one from
the CMakeLists.txt. The approach has to change because we removed
shared_sw, so we no longer have matching tags in our internal mirror and
github.

The other constraints for the implementation of client lib versioning
were: a) we want to improve debugability of customer issues by having
the exact git commit that they're using easily available in as many
cases as possible (to detect local changes) and b) we'll need to support
Python source distributions, so requiring git isn't an option.

To address a) this PR still records the git hash and dirty flag reported
by `git describe` and it's appended as build metadata to the full
version string. To address b) cmake works since the python source dist
has the CMakeLists.txt available. When a build is done from a source
distribution, we lose the ability to report the git hash, but that was
already the case and isn't a problem since we should only be releasing
source distributions of official releases.

This PR laso adds some compaitibility cmake find modules to avoid
conditional logic around supporting old platforms directly in the
CMakeLists.txt. The approach is to provide forward-compatible targets,
so the compatibility modules can just be dropped when we remove support
for 16.04. This PR also sets the default build type to "Release" using a
kitware-approved method.

Approved-by: Chris Bayruns
Approved-by: Pavlo Bashmakov