Add top-level CMakeLists.txt for ouster_example (MR !1734)
Rather than moving towards splitting up ouster_example into separate
repos, we'll be developing the basic tools for data capture and replay
as a single top-level project (the "Ouster SDK"). In this case, it makes
sense to build everything together, in a single (modular) cmake
project.

Building thie visualizer can be disabled by passing `-DBUILD_VIZ=No` to
cmake. This can be automated by detecting if dependency libraries are
present in the future.

This updates ros libraries and other code that was pulling in
ouster_client or ouster_viz separately. The ouster_example project will
be renamed when it's a bit more stable.

Approved-by: Chris Bayruns