Fix macos Python SDK sdist symlink issue (MR !1956)
Unlike on windows and linux, pip on macos chokes on the symlink to the
parent directory when building wheels. Traced system calls on the build
server to find that it was stuck in a symlink loop. This happens *only*
if there are existing builds in pip's tmpdir, so it works the first
time after a reboot.

Fix is to add a setuputils hook to prepare the source dir before
creating a sdist and clean up afterwards. Also added python3.6 to the
build while logged into the jenkins agent.

Approved-by: Kairen Wong
Approved-by: Chris Bayruns
Approved-by: Pavlo Bashmakov