Add release checklist and HIL tests (MR !2066)
Started out as a script to debug azimuth window issues, expanded to
cover some parts of the SDK that lack unit test coverage.

These tests currently aren't meant to be run in a totally automated way;
they intentionally check performance-dependent things like dropped packets
that would be unreliable on a shared build machine. There probably will
be spurious failures that need a human to interpret them or rerun the test.

Approved-by: Kairen Wong
Approved-by: Chris Bayruns
Approved-by: Pavlo Bashmakov