Add cmake install and packaging for ouster_example (MR !2124)
This adds support for the `install`, `package`, and `package_source`
targets to the ouster_example build. The `package` target will build a
distribution tarball and deb package, and is currently only supported on
linux. Files are currently installed under /opt/ouster, and can be used
in cmake with `find_package(OusterSDK)`.

Also add namespaced alias targets to make it easy to switch between an
installed SDK and building from source.

Approved-by: Kairen Wong
Approved-by: Chris Bayruns
Approved-by: Pavlo Bashmakov