Address MR comments from github push
Couple of bugs found in review in the case that the user specifies only
one port:

First bug was an incorrect assert: port guessing can return 0, meaning
no port was found for a lidar/imu stream that the user had specified.

Second bug was that a lone port guess could be returned even if the
sampled packets contain a valid matching data stream on a port other
than the one specified by the user. This was due to filtering on ports
before joining results on source ip.

* Fix bugs and add unit tests for those cases
* Some comment improvements
* Improve test fixtures for pcap unit tests