Update sdk viz to prefer glad instead of GLEW
Building against GLEW appears to cause some problems for manylinux
wheels, since it links directly to GLX and pulls in some additional
shared lib dependencies, which get included in the packages by
auditwheel. This is problematic for portability (see: the recent issue
with libsm, not to mention vendoring our own libGLX) and potentially
licensing due to accidentally redistributing lots of GPL libraries.

This updates the build to vcpkg 2022.02.23 and to prefer "glad," a
lighter-weight opengl loader library recommended by glfw and available
in vcpkg. GLEW is still supported for developer builds with just a
couple of #ifdefs.