Fix exported targets in sdk conan package
An earlier attempt to fix the SDK cmake config module broke it for the
conan package. Trying to use ${libtins_LIBRARIES} to deal with the
difference in target names between conan and regular cmake ended up
accidentally propagating hard-coded paths into OusterSDKConfig.cmake
under conan. This doesn't work because the location of dependency
libraries is different on each machine using conan.

This change adds a compatibility "interface" target for libtins that
matches the target name under conan. This lets the compatibility code
live in the Findlibtins.cmake module, so the interface to the
CMakeLists.txt is just a single "libtins" target. It also avoids having
to explicitly check for a conan environment via ${CONAN_EXPORTED}.

* Remove unused Findpcap module and associated find_package() call
* Add an interface target that allows linking either via discovered
  library paths or an imported target on platforms where
  libtinsConfig.cmake is installed
* Install the target so the same logic can be used when linking against
  a pre-built SDK
* Use a similar approach for incompatible jsoncpp targets