Use a single source for SDK version strings
Move the `ouster_client_VERSION` string to the top-level CMakeLists.txt
and renames it to `OusterSDK_VERSION`. This version string can replace
the current calendar date version scheme in the future. The version is
parsed out of the top-level CMakeLists.txt and used for the Python and
conan packages, to make it easier to keep them in sync.

Also, install the new top-level build info header under
`ouster/impl/build.h`. This makes it possible to build all of our C++
examples against an installed SDK package.