Update Viz camera together with other objects in draw
Move the `camera_front` update to happen together with
other objects (clouds, images, etc), so it's not detached
when user want's to move camera pose together with some
changes in the clouds.

The original motivation of leaving camera update to be
happen always even if draw() is taking slow seems not needed.
Or I can't understand the logic, which is also a possibility.
What I see is that camera updates will not appear on the screen
without draw() happening and the check/guard that prevents
the other objects update is exactly not finished draw(), so
fast-forwarding camera updates will not impact user experience
in the events of "slowness" because all updates will be slow
and delayed on the screen (including camera pose updates).

So moving camera updates in the draw() function after the
`front_changed` check didn't seem to add any lag anywhere
but resolved the "detached" updates issue.