PointViz: fix Label re-use
PointViz has data objects and GL objects dual that
keeps the GL viz states for them with a draw function.
GL objects are reused whenever it's appropriate
and thus in some cases they are in a state that is
different from the new added data object.

Bug appears when a set of Label are added and later
removed in some arbitrary order with add operation
follows in arbitrary order as well.

By making Label dirty on every add/re-add we ensure
that state of Label is properly updated to the GL
dual.