OusterSDK + Ouster CLI 2023 Q2 Release (#534)
OusterSDK + Ouster CLI 2023 Q2 Release
================================

With this release we are adding a bunch of things that we've used
internally at Ouster for some
time and we hope it will be useful to a broader set of users well.

Major things:

# Ouster CLI (Python)
Command Line Tools `ouster-cli` that combines basic and common actions
to work with sensors:
- Discover connected sensors (`ouster-cli discover`)
- Configure sensors (`ouster-cli source <SENSOR> config`)
- Get sensors metadata (`ouster-cli source <SENSOR> metadata`)
- Record data to a pcap with metadata json stored along (`ouster-cli
source <SENSOR> record`)
- Visualize data from sensor or pcap (`ouster-cli source [<SENSOR> |
<PCAP>] viz`)
- Pcap slice and convertion operations (`ouster-cli source <PCAP>
{info,slice,convert}`)
- Run slam to get trajectories and registered point clouds (`ouster-cli
source [<SENSOR> |
  <PCAP>] slam`)

# Ouster Mapping (Python)
Mapping utilities + tools to save registered point clouds from pcap or
running sensors that
accounts for sensor motion and deskews the cloud.

# Ouster OSF (C++/Python)
C++/Python library to save stream of LidarScans with metadata. Used by
mapping module to work with
scans + poses.

For more details please see the CHANGELOG.rst file.