FLEETSW-5814: fix ouster-cli discover race condition
This is an attempt at fixing the race condition described in the ticket. If the AsyncServiceDiscovery is "shutdown" in async_on_service_state_change or async_display_service_info, do not submit new futures to the asyncio executor.
Also, a bug is fixed with the warnings output.

Fix a test

Fix test

FLEETSW-5814: refactor and add a test to ensure RuntimeError isn't run when a new future is added to the executor after shutdown

Fix some flake errs

Troubleshooting CI issues

Add pytest-asyncio to enable a test

Troubleshooting CI issue