Rescale and add CalRef Versions of Each Palette
This changes the range of palettes that get very dark to the brightest 80% of the range.
Implements changes from: https://gitlab.com/ouster/software/ouster-sdk/-/merge_requests/357

It also adds calref versions of each palette that reach the top of their curve at 100% lambertian reflectivity that are used for reflectivity channel visualization only.

Closes FLEETSW-5324